INFUND-8747 ready fixed unit tests.

+++ ifs-web-service/ifs-application-service/src/main/java/org/innovateuk/ifs/application/populator/QuestionModelPopulator.java
@@ -46,8 +46,10 @@ public class QuestionModelPopulator extends BaseModelPopulator {
         }
 
         QuestionStatusResource questionStatusResource = questionService.getByQuestionIdAndApplicationIdAndOrganisationId(questionResource.getQuestion().getId(), questionResource.getApplication().getId(), questionResource.getCurrentApplicant().getOrganisation().getId());
-        List<QuestionStatusResource> notifications = questionService.getNotificationsForUser(asList(questionStatusResource), questionResource.getCurrentUser().getId());
-        questionService.removeNotifications(notifications);
+        if (questionStatusResource != null) {
+            List<QuestionStatusResource> notifications = questionService.getNotificationsForUser(asList(questionStatusResource), questionResource.getCurrentUser().getId());
+            questionService.removeNotifications(notifications);
+        }
     }
 
 }
